+++ packages/amplify-cli-core/src/index.ts
@@ -1,5 +1,6 @@
 import { ServiceSelection } from './serviceSelection';
 
+export * from './cfnUtilities';
 export * from './cliContext';
 export * from './cliContextEnvironmentProvider';
 export * from './cliEnvironmentProvider';
@@ -169,7 +170,7 @@ interface AmplifyToolkit {
   getPlugin: () => $TSAny;
   getCategoryPluginInfo: (context: $TSContext, category?: string, service?: string) => $TSAny;
   getAllCategoryPluginInfo: (context: $TSContext) => $TSAny;
-  getFrontendPlugins: () => $TSAny;
+  getFrontendPlugins: (context: $TSContext) => $TSAny;
   getEnvDetails: () => $TSAny;
   getEnvInfo: () => $TSAny;
   getProviderPlugins: (context: $TSContext) => $TSAny;