+++ packages/amplify-cli/src/commands/configure.ts
@@ -1,3 +1,4 @@
+import { $TSContext } from 'amplify-cli-core';
 import { analyzeProject } from '../config-steps/c0-analyzeProject';
 import { configFrontendHandler } from '../config-steps/c1-configFrontend';
 import { configProviders } from '../config-steps/c2-configProviders';
@@ -55,5 +56,5 @@ export const run = async (context: Context) => {
 
 function constructExeInfo(context: Context) {
   context.exeInfo = context.amplify.getProjectDetails();
-  context.exeInfo.inputParams = normalizeInputParams(context);
+  context.exeInfo.inputParams = normalizeInputParams((context as unknown) as $TSContext);
 }