+++ packages/amplify-provider-awscloudformation/src/pre-push-cfn-processor/cfn-pre-processor.ts
@@ -1,0 +1,26 @@
+import { pathManager, readCFNTemplate, writeCFNTemplate } from 'amplify-cli-core';
+import * as path from 'path';
+import { ProviderName as providerName } from '../constants';
+import { prePushCfnTemplateModifier } from './pre-push-cfn-modifier';
+
+const buildDir = 'build';
+
+/**
+ * Runs transformations on a CFN template and returns a path to the transformed template
+ *
+ * Expects to be run in an initialized Amplify project
+ * @param filePath the original template path
+ * @returns The file path of the modified template
+ */
+export async function preProcessCFNTemplate(filePath: string): Promise<string> {
+  const { templateFormat, cfnTemplate } = await readCFNTemplate(filePath);
+
+  await prePushCfnTemplateModifier(cfnTemplate);
+
+  const backendDir = pathManager.getBackendDirPath();
+  const pathSuffix = filePath.startsWith(backendDir) ? filePath.slice(backendDir.length) : path.parse(filePath).base;
+  const newPath = path.join(backendDir, providerName, buildDir, pathSuffix);
+
+  await writeCFNTemplate(cfnTemplate, newPath, { templateFormat });
+  return newPath;
+}