+++ packages/amplify-category-xr/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.8.3](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-xr@2.8.2...amplify-category-xr@2.8.3) (2021-04-19)
+
+**Note:** Version bump only for package amplify-category-xr
+
+
+
+
+
 ## [2.8.2](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-xr@2.8.1...amplify-category-xr@2.8.2) (2021-04-14)
 
 **Note:** Version bump only for package amplify-category-xr