+++ packages/amplify-nodejs-function-template-provider/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.6.3](https://github.com/aws-amplify/amplify-cli/compare/amplify-nodejs-function-template-provider@1.6.2...amplify-nodejs-function-template-provider@1.6.3) (2021-04-19)
+
+**Note:** Version bump only for package amplify-nodejs-function-template-provider
+
+
+
+
+
 ## [1.6.2](https://github.com/aws-amplify/amplify-cli/compare/amplify-nodejs-function-template-provider@1.6.1...amplify-nodejs-function-template-provider@1.6.2) (2021-04-14)
 
 **Note:** Version bump only for package amplify-nodejs-function-template-provider