+++ packages/amplify-python-function-runtime-provider/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.7.3](https://github.com/aws-amplify/amplify-cli/compare/amplify-python-function-runtime-provider@1.7.2...amplify-python-function-runtime-provider@1.7.3) (2021-04-19)
+
+**Note:** Version bump only for package amplify-python-function-runtime-provider
+
+
+
+
+
 ## [1.7.2](https://github.com/aws-amplify/amplify-cli/compare/amplify-python-function-runtime-provider@1.7.1...amplify-python-function-runtime-provider@1.7.2) (2021-04-14)
 
 **Note:** Version bump only for package amplify-python-function-runtime-provider